+++ CHANGELOG.md
@@ -10,6 +10,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Made rename operation fully backed by find all references, to ensure both return results that are consistent with each other
 
+### Fixed
+
+- Fixed Find All References / Rename not working on a table property defined inline, such as `name` in:
+
+```lua
+local T = {
+  name = "string"
+}
+```
+
 ## [1.26.0] - 2023-11-19
 
 ### Added